refactor: move shared components
