+++ web/src/beta/features/Visualizer/StoryPanel/Block/builtin/Camera/Content.tsx
@@ -5,7 +5,7 @@ import { useVisualizer } from "@reearth/beta/lib/core/Visualizer";
 import { useT } from "@reearth/services/i18n";
 import { styled } from "@reearth/services/theme";
 
-import { BlockContext } from "../../../../../../lib/core/shared/components/BlockWrapper";
+import { BlockContext } from "../../../../shared/components/BlockWrapper";
 
 import CameraEditor, { type CameraBlock as CameraBlockType } from "./Editor";
