+++ web/src/beta/features/Visualizer/StoryPanel/Block/builtin/Text/index.tsx
@@ -3,8 +3,8 @@ import { useMemo } from "react";
 import RichText from "@reearth/beta/lib/lexical/RichTextEditor";
 import { ValueTypes } from "@reearth/beta/utils/value";
 
-import BlockWrapper from "../../../../../../lib/core/shared/components/BlockWrapper";
-import { CommonBlockProps as BlockProps } from "../../../../../../lib/core/shared/types";
+import BlockWrapper from "../../../../shared/components/BlockWrapper";
+import { CommonBlockProps as BlockProps } from "../../../../shared/types";
 import { StoryBlock } from "../../../types";
 
 import TextBlockEditor from "./Editor";
