+++ web/src/beta/features/Visualizer/StoryPanel/Block/builtin/Timeline/Editor.tsx
@@ -1,7 +1,7 @@
 import Icon from "@reearth/beta/components/Icon";
 import * as Popover from "@reearth/beta/components/Popover";
+import useTimelineBlock from "@reearth/beta/features/Visualizer/shared/hooks/useTimelineBlock";
 import useHooks from "@reearth/beta/features/Visualizer/StoryPanel/Block/builtin/Timeline/hook";
-import useTimelineBlock from "@reearth/beta/lib/core/shared/hooks/useTimelineBlock";
 import { useT } from "@reearth/services/i18n";
 import { styled } from "@reearth/services/theme";
