+++ web/src/beta/features/Visualizer/StoryPanel/Block/builtin/Title/index.tsx
@@ -6,8 +6,8 @@ import { ValueTypes } from "@reearth/beta/utils/value";
 import { useT } from "@reearth/services/i18n";
 import { styled } from "@reearth/services/theme";
 
-import BlockWrapper from "../../../../../../lib/core/shared/components/BlockWrapper";
-import { CommonBlockProps as BlockProps } from "../../../../../../lib/core/shared/types";
+import BlockWrapper from "../../../../shared/components/BlockWrapper";
+import { CommonBlockProps as BlockProps } from "../../../../shared/types";
 import { StoryBlock } from "../../../types";
 
 const TitleBlock: React.FC<BlockProps<StoryBlock>> = ({ block, isSelected, ...props }) => {
