+++ web/src/beta/features/Visualizer/StoryPanel/Block/builtin/Video/index.tsx
@@ -1,8 +1,8 @@
 import { useMemo } from "react";
 
+import BlockWrapper from "@reearth/beta/features/Visualizer/shared/components/BlockWrapper";
+import type { CommonBlockProps as BlockProps } from "@reearth/beta/features/Visualizer/shared/types";
 import VideoPlayer from "@reearth/beta/features/Visualizer/StoryPanel/Block/builtin/Video/VideoPlayer";
-import BlockWrapper from "@reearth/beta/lib/core/shared/components/BlockWrapper";
-import type { CommonBlockProps as BlockProps } from "@reearth/beta/lib/core/shared/types";
 import type { ValueTypes } from "@reearth/beta/utils/value";
 
 import { StoryBlock } from "../../../types";
