+++ web/src/beta/features/Visualizer/shared/components/ContentWrapper/index.tsx
