+++ web/src/beta/features/Visualizer/shared/components/SelectableArea/ActionPanel/index.tsx
