+++ web/src/beta/features/Visualizer/shared/components/SelectableArea/hooks.ts
