+++ web/src/beta/features/Visualizer/useInfobox.ts
@@ -8,7 +8,6 @@ import { Infobox } from "./Crust/Infobox/types";
 export default function useInfobox({ layers }: { layers?: Layer[] }) {
   const { selectedLayer } = useContext(coreContext);
 
-  // Infobox
   const infobox: Infobox | undefined = useMemo(() => {
     if (!selectedLayer?.layer?.layer.infobox) return undefined;
     const selected = layers?.find(l => l.id === selectedLayer.layerId);
